+++ packages/transformers/js/core/src/utils.rs
@@ -13,8 +13,8 @@ use swc_core::common::errors::Emitter;
 use swc_core::ecma::ast::Ident;
 use swc_core::ecma::ast::IdentName;
 use swc_core::ecma::ast::{self};
-use swc_core::ecma::atoms::JsWord;
-use swc_core::ecma::atoms::js_word;
+use swc_core::ecma::atoms::Atom;
+use swc_core::ecma::atoms::atom;
 
 pub fn is_unresolved(ident: &Ident, unresolved_mark: Mark) -> bool {
   ident.ctxt.outer() == unresolved_mark
@@ -28,6 +28,7 @@ pub fn match_member_expr(expr: &ast::MemberExpr, idents: Vec<&str>, unresolved_m
 
   let mut member = expr;
   let mut idents = idents;
+
   while idents.len() > 1 {
     let expected = idents.pop().unwrap();
     let prop = match &member.prop {
@@ -61,7 +62,7 @@ pub fn match_member_expr(expr: &ast::MemberExpr, idents: Vec<&str>, unresolved_m
 }
 
 pub fn create_require(
-  specifier: swc_core::ecma::atoms::JsWord,
+  specifier: swc_core::ecma::atoms::Atom,
   unresolved_mark: Mark,
 ) -> ast::CallExpr {
   let mut normalized_specifier = specifier;
@@ -98,7 +99,7 @@ fn is_marked(mut ctxt: SyntaxContext, mark: Mark) -> bool {
   }
 }
 
-pub fn match_str(node: &ast::Expr) -> Option<(JsWord, Span)> {
+pub fn match_str(node: &ast::Expr) -> Option<(Atom, Span)> {
   use ast::*;
 
   match node {
@@ -112,7 +113,7 @@ pub fn match_str(node: &ast::Expr) -> Option<(JsWord, Span)> {
   }
 }
 
-pub fn match_property_name(node: &ast::MemberExpr) -> Option<(JsWord, Span)> {
+pub fn match_property_name(node: &ast::MemberExpr) -> Option<(Atom, Span)> {
   match &node.prop {
     ast::MemberProp::Computed(s) => match_str(&s.expr),
     ast::MemberProp::Ident(id) => Some((id.sym.clone(), id.span)),
@@ -120,7 +121,7 @@ pub fn match_property_name(node: &ast::MemberExpr) -> Option<(JsWord, Span)> {
   }
 }
 
-pub fn match_export_name(name: &ast::ModuleExportName) -> (JsWord, Span) {
+pub fn match_export_name(name: &ast::ModuleExportName) -> (Atom, Span) {
   match name {
     ast::ModuleExportName::Ident(id) => (id.sym.clone(), id.span),
     ast::ModuleExportName::Str(s) => (s.value.clone(), s.span),
@@ -135,14 +136,14 @@ pub fn match_export_name_ident(name: &ast::ModuleExportName) -> &ast::Ident {
   }
 }
 
-pub fn match_require(node: &ast::Expr, unresolved_mark: Mark, ignore_mark: Mark) -> Option<JsWord> {
+pub fn match_require(node: &ast::Expr, unresolved_mark: Mark, ignore_mark: Mark) -> Option<Atom> {
   use ast::*;
 
   match node {
     Expr::Call(call) => match &call.callee {
       Callee::Expr(expr) => match &**expr {
         Expr::Ident(ident) => {
-          if ident.sym == js_word!("require")
+          if ident.sym == atom!("require")
             && is_unresolved(ident, unresolved_mark)
             && !is_marked(ident.ctxt, ignore_mark)
             && let Some(arg) = call.args.first()
@@ -169,7 +170,7 @@ pub fn match_require(node: &ast::Expr, unresolved_mark: Mark, ignore_mark: Mark)
   }
 }
 
-pub fn match_import(node: &ast::Expr) -> Option<JsWord> {
+pub fn match_import(node: &ast::Expr) -> Option<Atom> {
   use ast::*;
 
   match node {
@@ -188,14 +189,14 @@ pub fn match_import(node: &ast::Expr) -> Option<JsWord> {
 
 /// This matches an expression like `importCond('if_true_dependency_id`, 'if_false_dependency_id')` and
 /// returns the two dependency ids.
-pub fn match_import_cond(node: &ast::Expr, ignore_mark: Mark) -> Option<(JsWord, JsWord)> {
+pub fn match_import_cond(node: &ast::Expr, ignore_mark: Mark) -> Option<(Atom, Atom)> {
   use ast::*;
 
   match node {
     Expr::Call(call) => match &call.callee {
       Callee::Expr(expr) => match &**expr {
         Expr::Ident(ident) => {
-          if ident.sym == js_word!("importCond")
+          if ident.sym == atom!("importCond")
             && !is_marked(ident.ctxt, ignore_mark)
             && call.args.len() == 2
           {
@@ -218,7 +219,7 @@ pub fn match_import_cond(node: &ast::Expr, ignore_mark: Mark) -> Option<(JsWord,
 
 // `name` must not be an existing binding.
 pub fn create_global_decl_stmt(
-  name: swc_core::ecma::atoms::JsWord,
+  name: swc_core::ecma::atoms::Atom,
   init: ast::Expr,
   global_mark: Mark,
 ) -> (ast::Stmt, SyntaxContext) {
@@ -247,7 +248,7 @@ pub fn create_global_decl_stmt(
 
 pub fn get_undefined_ident(unresolved_mark: Mark) -> ast::Ident {
   ast::Ident::new(
-    js_word!("undefined"),
+    atom!("undefined"),
     DUMMY_SP,
     SyntaxContext::empty().apply_mark(unresolved_mark),
   )
@@ -435,7 +436,7 @@ pub struct ErrorBuffer(
 );
 
 impl Emitter for ErrorBuffer {
-  fn emit(&mut self, db: &DiagnosticBuilder) {
+  fn emit(&mut self, db: &mut DiagnosticBuilder) {
     self.0.lock().push((**db).clone());
   }
 }