+++ packages/core/integration-tests/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @atlaspack/integration-tests
 
+## 2.20.1
+
+### Patch Changes
+
+- [#942](https://github.com/atlassian-labs/atlaspack/pull/942) [`695339e`](https://github.com/atlassian-labs/atlaspack/commit/695339e4fd46df9590013c2ef4de74cb22225927) Thanks [@marcins](https://github.com/marcins)! - Improve error handling in Tokens transform
+
+- [#943](https://github.com/atlassian-labs/atlaspack/pull/943) [`6152999`](https://github.com/atlassian-labs/atlaspack/commit/61529999f8ab3131c591362c88b65e697e0bd9b8) Thanks [@emzeidan](https://github.com/emzeidan)! - Update TesseractResolver.ts to add new `unsupportedExtensions` config option to allow fallback
+
 ## 2.20.0
 
 ### Minor Changes